Zero layer thickness layers are considered active due to float rounding (#324)
* Zero layer thickness layers are considered active due to float rounding

Subtracting two botms to calculate the thickness results is some layers with a thickness of ~1e-9. This rounding later results in a problem because this positive thickness leads to an idomain of 1.

This can either be solved in `nlmod.layers.calculating_thickness()` or in `nlmod.layers.get_idomain()`. What do you prefer?

* Remove trailing whitespace

* Applied patch also to get_idomain